% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range01.R
\name{range01}
\alias{range01}
\title{Standardize variables 0 to 1}
\usage{
range01(x)
}
\description{
This function will standardize your variables from 0 to 1.
}
\examples{
range01()
}
\keyword{standardization}
