% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_names.R
\name{var_names}
\alias{var_names}
\title{A Cat Function}
\usage{
var_names(data, keyword = "")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
var_names()
}
\keyword{cats}
